Logic Conditions Exercises
Most interesting thing learned in this module so far is use of the Ternary operator. 

Short hand version of an if else statement in one line. 

Example here, we have the condition to check logic on, followed by a question mark, then if the true and after the colon we see the false response. 
string response = (permission.Contains("Admin") || level >= 55 ? "Welcome, Super Admin User" : "Not a super user");